separate widgets tests
